backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Pillow==9.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.6
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
tzdata==2022.6
whitenoise==5.3.0